Incorporate the OrderManager package. (#3)
* Incorporate the OrderManager package.

This is the key logic for handling the certificate ordering process.  The logic is based heavily on Go's autocert package, but reoriented into a form that is more amenable to a reconciliation loop.

I have not found any good fakes for the acme package that would allow for this package to be reasonably unit tested, so instead I wrote a very simple standalone sample that can be used to exercise it's functionality in a manner similar to autocert.

I have stood this up temporarily [here](https://foo.docs-on-the-rocks.io) to confirm that nothing broke in the copy.

* Incorporate review comments from @tcnghia